docs/requirements.txt
python-fasthtml
fh_frankenui @ git+https://github.com/isaac-flath/fh-frankenui.git